feat:added search and filtering for places

+++ Backend/src/Controllers/Users/governor.controller.ts
@@ -1,9 +1,10 @@
 import { Request, Response } from "express";
+import { Types } from "mongoose";
 
 import { Governor } from "../../Database/Models/Users/governor.model";
 import { Places } from "../../Database/Models/places.model";
 
-export const crteatGovernor = async (req: Request, res: Response) => {
+export const createGovernor = async (req: Request, res: Response) => {
 	try {
 		const { username, email, password } = req.body;
 		if (!username || !email || !password) {
@@ -30,23 +31,43 @@ export const getGovernors = async (req: Request, res: Response) => {
 export const getHistoricalLocations = async (req: Request, res: Response) => {
 	try {
 		const { id } = req.params;
+
+		if (!id) {
+			return res.status(400).send("id is required");
+		}
+
+		if (!Types.ObjectId.isValid(id)) {
+			return res.status(400).send("id is invalid");
+		}
 		const historicalLocations = await Places.find({ governorId: id });
+
 		res.status(200).json(historicalLocations);
 	} catch (error) {
 		console.log(error);
-		res.status(500).send("error deleting Governor");
+		res.status(500).send("Error deleting Governor");
 	}
 };
 
 export const deleteGovernor = async (req: Request, res: Response) => {
 	try {
 		const { id } = req.params;
+
+		if (!id) {
+			return res.status(400).send("Id is Required");
+		}
+
+		if (!Types.ObjectId.isValid(id)) {
+			return res.status(400).send("Id is Invalid");
+		}
+
 		const governor = await Governor.findByIdAndDelete(id);
+
 		if (!governor) {
 			return res.status(404).send("governor not found");
 		}
-		res.status(200).send("deleted successfully");
+
+		res.status(200).send("Deleted successfully");
 	} catch (error) {
-		res.status(500).send("error deleting Governor");
+		res.status(500).send("Error while deleting Governor");
 	}
 };
